Refactor, move old auth functionality into separate file (authentication_direct_proxy.py)
